splitup tetst for BaseModel and DefaultModel
