+++ src/xcode/ENA/ENA/Source/Scenes/DataDonation/Model/DataDonationModel.swift
@@ -32,8 +32,6 @@ struct DataDonationModel {
 		}?.districtName
 	}
 
-	// MARK: - Public
-
 	// MARK: - Internal
 
 	var isConsentGiven: Bool
